Support filename associations
Adds an extra components page to the installer, on which users can choose
which file name associations to add to their system. These are pulled from
the installed shared-mime-info database and the launcher's source .desktop
files. Derivative types are shown as optional (and unchecked) components.
File types listed directly by the .desktop file are treated as "more
native" to the resultant app, and are checked by default.

Closes achadwick/styrene#1